geolocation/geo-00013.html

<!DOCTYPE html>
<html>
  <head>
    <title>Geolocation Test: call getCurrentPosition with null success and error callbacks, exception is thrown</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#get-current-position">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var t =async_test("Test checks that call getCurrentPosition with null success and error callbacks, exception is thrown");
      setup({timeout: 20000});
      try {
        window.navigator.geolocation.getCurrentPosition(null, null);
      } catch (err) {
        t.step(function () {
          assert_true(true, "call getCurrentPosition with null success and error callbacks, exception is thrown");
        });
        t.done();
      }
    </script>
  </body>
</html>
